refactor: moved files to settings directory
